+++ HalpBot-Core/src/main/java/nz/pumbas/halpbot/converters/ConverterServicePreProcessor.java
@@ -9,11 +9,11 @@ import org.dockbox.hartshorn.core.services.ServicePreProcessor;
 import nz.pumbas.halpbot.commands.annotations.UseCommands;
 
 @AutomaticActivation
-public class ConverterServicePreProcessor implements ServicePreProcessor<UseCommands>
+public class ConverterServicePreProcessor implements ServicePreProcessor<UseConverters>
 {
     @Override
-    public Class<UseCommands> activator() {
-        return UseCommands.class;
+    public Class<UseConverters> activator() {
+        return UseConverters.class;
     }
 
     @Override
