+++ HalpBot-Core/src/main/java/nz/pumbas/halpbot/converters/UseConverters.java
@@ -1,4 +1,4 @@
-package nz.pumbas.halpbot.commands.annotations;
+package nz.pumbas.halpbot.converters;
 
 import org.dockbox.hartshorn.core.annotations.activate.ServiceActivator;
 
@@ -7,14 +7,9 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import nz.pumbas.halpbot.common.UseDefault;
-import nz.pumbas.halpbot.converters.UseConverters;
-
-@UseDefault
-@UseConverters
 @ServiceActivator(scanPackages = "nz.pumbas.halpbot")
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface UseCommands
+public @interface UseConverters
 {
 }
