+++ HalpBot-Core/src/test/java/nz/pumbas/halpbot/commands/objects/Matrix.java
@@ -36,7 +36,7 @@ import nz.pumbas.halpbot.commands.exceptions.ErrorMessageException;
 import nz.pumbas.halpbot.commands.exceptions.UnimplementedFeatureException;
 import nz.pumbas.halpbot.converters.annotations.parameter.Implicit;
 
-@Service
+@Service(singleton = false)
 public class Matrix
 {
     public static final Matrix UnitSquare = new Matrix(2, 4, 0, 0, 1, 1, 0, 1, 0, 1);
